reorganize core upstream of smithy-client
