+++ .changeset/unlucky-books-think.md
@@ -1,0 +1,8 @@
+---
+"@smithy/middleware-compression": patch
+"@smithy/middleware-endpoint": patch
+"@smithy/smithy-client": patch
+"@smithy/core": patch
+---
+
+reorganize smithy/core to be upstream of smithy/smithy-client
