+++ packages/core/src/index.ts
@@ -1,8 +1,11 @@
+export * from "./getSmithyContext";
 export * from "./middleware-http-auth-scheme";
 export * from "./middleware-http-signing";
-export * from "./util-identity-and-auth";
-export * from "./getSmithyContext";
 export * from "./normalizeProvider";
-export * from "./protocols/requestBuilder";
-export * from "./setFeature";
 export { createPaginator } from "./pagination/createPaginator";
+export * from "./protocols/collect-stream-body";
+export * from "./protocols/requestBuilder";
+export * from "./protocols/resolve-path";
+export * from "./protocols/extended-encode-uri-component";
+export * from "./setFeature";
+export * from "./util-identity-and-auth";
