+++ packages/core/src/protocols/extended-encode-uri-component.spec.ts
