+++ packages/smithy-client/package.json
@@ -23,6 +23,7 @@
   },
   "license": "Apache-2.0",
   "dependencies": {
+    "@smithy/core": "workspace:^",
     "@smithy/middleware-endpoint": "workspace:^",
     "@smithy/middleware-stack": "workspace:^",
     "@smithy/protocol-http": "workspace:^",
