+++ packages/smithy-client/src/extended-encode-uri-component.ts
@@ -1,11 +1,5 @@
 /**
  * @internal
- *
- * Function that wraps encodeURIComponent to encode additional characters
- * to fully adhere to RFC 3986.
+ * Backwards compatibility re-export.
  */
-export function extendedEncodeURIComponent(str: string): string {
-  return encodeURIComponent(str).replace(/[!'()*]/g, function (c) {
-    return "%" + c.charCodeAt(0).toString(16).toUpperCase();
-  });
-}
+export { extendedEncodeURIComponent } from "@smithy/core";
